Factor out `transformERC20` decoding code to `test_utils/decoders.ts`
